Move: 에디터 페이지 및 컴포넌트 폴더 구조 변경
