+++ src/pages/editor/components/editor/extensions.ts
@@ -1,4 +1,4 @@
-import Title from '@/components/editor/Title';
+import Title from '@/pages/editor/components/editor/Title';
 import Color from '@tiptap/extension-color';
 import Document from '@tiptap/extension-document';
 import Heading from '@tiptap/extension-heading';
