+++ src/pages/editor/components/editor/styles.scss
