+++ src/pages/editor/edit/index.tsx
@@ -1,0 +1,12 @@
+import { Suspense } from 'react';
+import EditEditor from './components/EditEditor';
+
+const EditPage = () => {
+  return (
+    <Suspense>
+      <EditEditor />
+    </Suspense>
+  );
+};
+
+export default EditPage;
