+++ src/pages/editor/write/index.tsx
@@ -1,0 +1,7 @@
+import DefaultEditor from '@/pages/editor/components/editor';
+
+const WritePage = () => {
+  return <DefaultEditor />;
+};
+
+export default WritePage;
